rename: svg name change
